+++ Sources/Protobuf/room.pb.swift
@@ -472,6 +472,15 @@ struct RoomState {
 
   var visibility: Visibility = .public
 
+  var group: RoomState.Group {
+    get {return _group ?? RoomState.Group()}
+    set {_group = newValue}
+  }
+  /// Returns true if `group` has been explicitly set.
+  var hasGroup: Bool {return self._group != nil}
+  /// Clears the value of `group`. Subsequent reads from it will return its default value.
+  mutating func clearGroup() {self._group = nil}
+
   var unknownFields = SwiftProtobuf.UnknownStorage()
 
   struct RoomMember {
@@ -496,7 +505,25 @@ struct RoomState {
     init() {}
   }
 
+  struct Group {
+    // SwiftProtobuf.Message conformance is added in an extension below. See the
+    // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
+    // methods supported on all messages.
+
+    var id: Int64 = 0
+
+    var name: String = String()
+
+    var image: String = String()
+
+    var unknownFields = SwiftProtobuf.UnknownStorage()
+
+    init() {}
+  }
+
   init() {}
+
+  fileprivate var _group: RoomState.Group? = nil
 }
 
 struct CreateRequest {
@@ -507,6 +534,8 @@ struct CreateRequest {
   var name: String = String()
 
   var visibility: Visibility = .public
+
+  var group: Int64 = 0
 
   var unknownFields = SwiftProtobuf.UnknownStorage()
 
@@ -1037,6 +1066,7 @@ extension RoomState: SwiftProtobuf.Message, SwiftProtobuf._MessageImplementation
     3: .same(proto: "members"),
     4: .same(proto: "role"),
     5: .same(proto: "visibility"),
+    6: .same(proto: "group"),
   ]
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
@@ -1047,6 +1077,7 @@ extension RoomState: SwiftProtobuf.Message, SwiftProtobuf._MessageImplementation
       case 3: try decoder.decodeRepeatedMessageField(value: &self.members)
       case 4: try decoder.decodeSingularStringField(value: &self.role)
       case 5: try decoder.decodeSingularEnumField(value: &self.visibility)
+      case 6: try decoder.decodeSingularMessageField(value: &self._group)
       default: break
       }
     }
@@ -1068,6 +1099,9 @@ extension RoomState: SwiftProtobuf.Message, SwiftProtobuf._MessageImplementation
     if self.visibility != .public {
       try visitor.visitSingularEnumField(value: self.visibility, fieldNumber: 5)
     }
+    if let v = self._group {
+      try visitor.visitSingularMessageField(value: v, fieldNumber: 6)
+    }
     try unknownFields.traverse(visitor: &visitor)
   }
 
@@ -1077,6 +1111,7 @@ extension RoomState: SwiftProtobuf.Message, SwiftProtobuf._MessageImplementation
     if lhs.members != rhs.members {return false}
     if lhs.role != rhs.role {return false}
     if lhs.visibility != rhs.visibility {return false}
+    if lhs._group != rhs._group {return false}
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }
@@ -1141,11 +1176,53 @@ extension RoomState.RoomMember: SwiftProtobuf.Message, SwiftProtobuf._MessageImp
   }
 }
 
+extension RoomState.Group: SwiftProtobuf.Message, SwiftProtobuf._MessageImplementationBase, SwiftProtobuf._ProtoNameProviding {
+  static let protoMessageName: String = RoomState.protoMessageName + ".Group"
+  static let _protobuf_nameMap: SwiftProtobuf._NameMap = [
+    1: .same(proto: "id"),
+    2: .same(proto: "name"),
+    3: .same(proto: "image"),
+  ]
+
+  mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
+    while let fieldNumber = try decoder.nextFieldNumber() {
+      switch fieldNumber {
+      case 1: try decoder.decodeSingularInt64Field(value: &self.id)
+      case 2: try decoder.decodeSingularStringField(value: &self.name)
+      case 3: try decoder.decodeSingularStringField(value: &self.image)
+      default: break
+      }
+    }
+  }
+
+  func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
+    if self.id != 0 {
+      try visitor.visitSingularInt64Field(value: self.id, fieldNumber: 1)
+    }
+    if !self.name.isEmpty {
+      try visitor.visitSingularStringField(value: self.name, fieldNumber: 2)
+    }
+    if !self.image.isEmpty {
+      try visitor.visitSingularStringField(value: self.image, fieldNumber: 3)
+    }
+    try unknownFields.traverse(visitor: &visitor)
+  }
+
+  static func ==(lhs: RoomState.Group, rhs: RoomState.Group) -> Bool {
+    if lhs.id != rhs.id {return false}
+    if lhs.name != rhs.name {return false}
+    if lhs.image != rhs.image {return false}
+    if lhs.unknownFields != rhs.unknownFields {return false}
+    return true
+  }
+}
+
 extension CreateRequest: SwiftProtobuf.Message, SwiftProtobuf._MessageImplementationBase, SwiftProtobuf._ProtoNameProviding {
   static let protoMessageName: String = "CreateRequest"
   static let _protobuf_nameMap: SwiftProtobuf._NameMap = [
     1: .same(proto: "name"),
     2: .same(proto: "visibility"),
+    3: .same(proto: "group"),
   ]
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
@@ -1153,6 +1230,7 @@ extension CreateRequest: SwiftProtobuf.Message, SwiftProtobuf._MessageImplementa
       switch fieldNumber {
       case 1: try decoder.decodeSingularStringField(value: &self.name)
       case 2: try decoder.decodeSingularEnumField(value: &self.visibility)
+      case 3: try decoder.decodeSingularInt64Field(value: &self.group)
       default: break
       }
     }
@@ -1165,12 +1243,16 @@ extension CreateRequest: SwiftProtobuf.Message, SwiftProtobuf._MessageImplementa
     if self.visibility != .public {
       try visitor.visitSingularEnumField(value: self.visibility, fieldNumber: 2)
     }
+    if self.group != 0 {
+      try visitor.visitSingularInt64Field(value: self.group, fieldNumber: 3)
+    }
     try unknownFields.traverse(visitor: &visitor)
   }
 
   static func ==(lhs: CreateRequest, rhs: CreateRequest) -> Bool {
     if lhs.name != rhs.name {return false}
     if lhs.visibility != rhs.visibility {return false}
+    if lhs.group != rhs.group {return false}
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }